+++ app/sprinkles/admin/routes/admin.php
@@ -10,15 +10,15 @@
 /**
  * Routes for administrative panel management.
  */
-$app->group('/admin', function () {
+$app->group('/dashboard', function () {
     $this->get('', 'UserFrosting\Sprinkle\Admin\Controller\AdminController:pageDashboard')
          ->setName('dashboard');
 })->add('authGuard');
 
-$app->group('/api/admin', function () {
+$app->group('/api/dashboard', function () {
     $this->post('/clear-cache', 'UserFrosting\Sprinkle\Admin\Controller\AdminController:clearCache');
 })->add('authGuard');
 
-$app->group('/modals/admin', function () {
+$app->group('/modals/dashboard', function () {
     $this->get('/clear-cache', 'UserFrosting\Sprinkle\Admin\Controller\AdminController:getModalConfirmClearCache');
 })->add('authGuard');